Revert "Merge latest from the official repo"
This reverts commit 6c56e7d726f4637883087cd680db70fd131e93a2.
